Load JSRoot once and use that
